# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.50
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/memory/l3:
    get:
      tags: [Memory]
      summary: List L3 distilled personas
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: includeDeleted
          in: query
          schema: { type: string, enum: [active, deleted, any], default: active }
      responses:
        "200":
          description: Paginated L3 personas
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/L3Working" }
                  pagination: { $ref: "#/components/schemas/MemoryPagination" }
        "400": { description: Invalid query parameters }
    post:
      tags: [Memory]
      summary: Enqueue global L3 regeneration
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryRegenerateBody" }
      responses:
        "200": { description: Regeneration task enqueued }
        "400": { description: Validation or storage error }
        "409": { description: Rolling-window error cap reached }
...
  /api/memory/pipeline-settings:
    get:
      tags: [Memory]
      summary: Read owner-scoped memory capture and injection settings
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Effective per-key settings with source layer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/MemoryPipelineSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Cross-owner settings access denied }
    put:
      tags: [Memory]
      summary: Set owner-scoped memory capture and injection settings
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [captureEnabled, injectionEnabled]
              properties:
                captureEnabled: { type: boolean }
                injectionEnabled: { type: boolean }
              additionalProperties: false
      responses:
        "200": { description: Per-key settings stored }
        "400": { description: Invalid settings body }
        "403": { description: Cross-owner settings access denied }
    delete:
      tags: [Memory]
      summary: Reset owner-scoped settings to environment or default fallback
      security:
        - BearerAuth: []
        - ManagementSessionAuth: []
      parameters:
        - name: apiKeyId
          in: query
          schema: { type: string }
      responses:
        "200": { description: Per-key override removed and fallback returned }
        "403": { description: Cross-owner settings access denied }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the OmniRoute dashboard
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie for protected management routes
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
...
  schemas:
...
    MemoryPagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    MemoryRegenerateBody:
      type: object
      properties:
        reason: { type: string, maxLength: 200 }
...
    L3Working:
      type: object
      description: The owner-visible distilled L3 persona
      required:
        [
          id,
          ownerApiKeyId,
          content,
          promptMode,
          version,
          createdAt,
          updatedAt,
          lastModifiedBy,
          editedByUser,
        ]
      properties:
        id: { type: string }
        ownerApiKeyId: { type: string }
        content: { type: string }
        promptMode: { type: string, enum: [chat, code] }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastModifiedBy: { type: string, enum: [user, pipeline] }
        editedByUser: { type: boolean }
        deletedAt: { type: [string, "null"], format: date-time }
...
    MemoryPipelineSettings:
      type: object
      required:
        [
          captureEnabled,
          injectionEnabled,
          sourceLayer,
          apiKeyId,
          l3CharBudget,
          l2CharBudget,
          l1CharBudget,
          totalCharBudget,
          recallTimeoutMs,
        ]
      properties:
        captureEnabled: { type: boolean }
        injectionEnabled: { type: boolean }
        sourceLayer: { type: string, enum: [per-key, env, default] }
        apiKeyId: { type: [string, "null"] }
        l3CharBudget: { type: integer, minimum: 0, maximum: 64000 }
        l2CharBudget: { type: integer, minimum: 0, maximum: 64000 }
        l1CharBudget: { type: integer, minimum: 0, maximum: 64000 }
        totalCharBudget: { type: integer, minimum: 0, maximum: 64000 }
        recallTimeoutMs: { type: integer, minimum: 1, maximum: 60000 }
...